refactor methods and attributes to ops_task.py
