+++ backend/data_tools/src/pipeline_data_from_excel/ops_task.py
@@ -1,42 +1,18 @@
-import luigi
-from models import *
-from sqlalchemy.future import create_engine
-from sqlalchemy.orm import Session
+import logging
 
-engine = create_engine(
-    "postgresql://postgres:local_password@localhost:5432/postgres"  # pragma: allowlist secret
-)
+import luigi
+from sqlalchemy.future import create_engine
 
 
 class OPSTask(luigi.Task):
     def __init__(self, run_date):
-        self._engine = engine
+        self.engine = create_engine(
+            "postgresql://postgres:local_password@localhost:5432/postgres"  # pragma: allowlist secret
+        )
+        self.logger = logging.getLogger("luigi-interface")
+        self.task_meta = {"log_messages": []}
         super().__init__(run_date)
 
-
-@OPSTask.event_handler(luigi.Event.SUCCESS)
-def success(task):
-    with Session(engine) as session, session.begin():
-        session.add(
-            ETLTaskStatus(
-                workflow_name="etl_data_from_excel",
-                task_name=task.task_module,
-                run_at=task.run_date,
-                task_meta=task.task_meta,
-                status="SUCCESS",
-            )
-        )
-
-
-@OPSTask.event_handler(luigi.Event.FAILURE)
-def fail(task):
-    with Session(engine) as session, session.begin():
-        session.add(
-            ETLTaskStatus(
-                workflow_name="etl_data_from_excel",
-                task_name=task.task_module,
-                run_at=task.run_date,
-                task_meta=task.task_meta,
-                status="FAIL",
-            )
-        )
+    def _log_message(self, message):
+        self.logger.info(message)
+        self.task_meta["log_messages"].append(message)
