+++ backend/data_tools/src/pipeline_data_from_excel/pipeline_business_rules.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from typing import List, TypeVar
+from typing import List, Optional, TypeVar
 
 SourceEntity = TypeVar("SourceTable")
 DestinationEntity = TypeVar("DestinationEntity")
@@ -9,6 +9,6 @@ class PipelineBusinessRules(ABC):
     @staticmethod
     def apply_business_rules(
         data: List[SourceEntity],
-        existing_data: List[DestinationEntity] = [],
+        existing_data: Optional[List[DestinationEntity]] = [],
     ) -> List[DestinationEntity]:
         ...
